+++ src/components/common/View/LoadingView.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { HEADER_HEIGHT } from '@/components/features/Layout/Header';
+import { HEADER_HEIGHT } from '@/components/Layout/Header';
 
 import { Spinner } from '../Spinner';
