+++ src/components/features/Order/OrderForm/index.tsx
@@ -5,7 +5,7 @@ import { Spacing } from '@/components/common/layouts/Spacing';
 import { SplitLayout } from '@/components/common/layouts/SplitLayout';
 import type { OrderFormData, OrderHistory } from '@/types';
 
-import { HEADER_HEIGHT } from '../../Layout/Header';
+import { HEADER_HEIGHT } from '../../../Layout/Header';
 import { GoodsInfo } from './GoodsInfo';
 import { OrderFormMessageCard } from './MessageCard';
 import { OrderFormOrderInfo } from './OrderInfo';
